AddTaskScreen refactored to smaller containers and own packages
