+++ app/src/main/java/com/metropolia/eatthefrog/constants/Constants.kt
@@ -2,3 +2,4 @@ package com.metropolia.eatthefrog.constants
 
 const val DATE_FORMAT = "dd.MM.yyyy"
 const val TIME_FORMAT = "HH:mm"
+const val MUX_SUBTASK_AMOUNT = 7
