+++ app/src/main/java/com/metropolia/eatthefrog/navigation/Navigation.kt
@@ -7,9 +7,9 @@ import androidx.compose.runtime.MutableState
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.metropolia.eatthefrog.screens.AddTaskScreen
 import com.metropolia.eatthefrog.screens.HomeScreen
 import com.metropolia.eatthefrog.screens.ProfileScreen
+import com.metropolia.eatthefrog.screens.addTask.AddTaskScreen
 
 @ExperimentalMaterialApi
 @Composable
